+++ src/components/molecules/index.ts
@@ -14,7 +14,7 @@ import Profile from "@/components/molecules/profile/Profile";
 import Rating from "@/components/molecules/rating/Rating";
 import ReviewForm from "@/components/molecules/review-form/ReviewForm";
 import SearchBar from "@/components/molecules/searchBar/SearchBar";
-import ToastContainer from "@/components/molecules/toast/ToastContainer";
+import ToastContainer from "@/components/molecules/toast-container/ToastContainer";
 
 export {
 	QuillEditor,
